requirements.txt
openai-agents==0.9.3
sympy>=1.13
rich>=13.0
httpx>=0.27
ddgs>=9.0
